enable manually modified table.xlsx

+++ main.py
@@ -10,11 +10,14 @@ folder = os.getcwd() + '/downloads/'
 
 if not os.path.exists(folder):
     os.mkdir(folder)
+    
+if not os.path.exists("table.xlsx")
+    books = pd.read_excel('https://resource-cms.springernature.com/springer-cms/rest/v1/content/17858272/data/v4')
 
-books = pd.read_excel('https://resource-cms.springernature.com/springer-cms/rest/v1/content/17858272/data/v4')
-
-# save table:
-books.to_excel(folder + 'table.xlsx')
+    # save table:
+    books.to_excel(folder + 'table.xlsx')
+else:
+    books = pd.read_excel('table.xlsx', index_col=None, header=None)  
 
 # debug:
 # books = books.head()
